+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## upcoming release
+- Add support for --squash experimental build option ([248][])
+
+[248]: https://github.com/spotify/dockerfile-maven/pull/248
+
 ## 1.4.9 (released October 25 2018)
 - Upgrade docker-client dep from 8.14.2 to 8.14.3 to fix spotify/docker-client#1100